Factorize cache events.
