--- src/Illuminate/Cache/Events/CacheEvent.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Illuminate\Cache\Events;
-
-class CacheEvent
-{
-    /**
-     * Set the tags for the cache event.
-     *
-     * @param  array  $tags
-     * @return $this
-     */
-    public function setTags($tags)
-    {
-        $this->tags = $tags;
-
-        return $this;
-    }
-}
