+++ src/Illuminate/Cache/Events/CacheHit.php
@@ -2,28 +2,14 @@
 
 namespace Illuminate\Cache\Events;
 
-class CacheHit extends CacheEvent
+class CacheHit extends AbstractCacheEvent
 {
-    /**
-     * The key that was hit.
-     *
-     * @var string
-     */
-    public $key;
-
     /**
      * The value that was retrieved.
      *
      * @var mixed
      */
     public $value;
-
-    /**
-     * The tags that were assigned to the key.
-     *
-     * @var array
-     */
-    public $tags;
 
     /**
      * Create a new event instance.
@@ -35,8 +21,7 @@ class CacheHit extends CacheEvent
      */
     public function __construct($key, $value, array $tags = [])
     {
-        $this->key = $key;
-        $this->tags = $tags;
+        parent::__construct($key, $tags);
         $this->value = $value;
     }
 }
