+++ src/Illuminate/Cache/Events/CacheMissed.php
@@ -2,32 +2,7 @@
 
 namespace Illuminate\Cache\Events;
 
-class CacheMissed extends CacheEvent
+class CacheMissed extends AbstractCacheEvent
 {
-    /**
-     * The key that was missed.
-     *
-     * @var string
-     */
-    public $key;
-
-    /**
-     * The tags that were assigned to the key.
-     *
-     * @var array
-     */
-    public $tags;
-
-    /**
-     * Create a new event instance.
-     *
-     * @param  string  $key
-     * @param  array  $tags
-     * @return void
-     */
-    public function __construct($key, array $tags = [])
-    {
-        $this->key = $key;
-        $this->tags = $tags;
-    }
+    //
 }
